+++ src/interfaces/options/ITempCanvasOptions.ts
@@ -4,4 +4,6 @@ import ISimpleSize from '../simple/ISimpleSize'
 export default interface ITempCanvasOptions {
     size: ISimpleSize
     callback: (ctx: Context) => void
+    quality?: number
+    mimeImageType?: string
 }