+++ src/objects/shapeable/PointerableObject.ts
@@ -1,33 +1,29 @@
-import IRectangle from '../interfaces/IRectangle'
-import IShapeConfig from '../interfaces/IShapeConfig'
-import Color from '../utils/Color'
-import { Context, PointerCallback } from '../utils/types'
+import IRectangle from '../../interfaces/IRectangle'
+import IShapeConfig from '../../interfaces/IShapeConfig'
+import Color from '../../utils/Color'
+import { Context, PointCallback } from '../../utils/types'
 import ShapeableObject from './ShapeableObject'
 
 export default class PointerableObject extends ShapeableObject implements IShapeConfig, IRectangle {
-    private _nonDownCallbackFactory(cb: PointerCallback): PointerCallback {
+    private _nonDownCallbackFactory(cb: PointCallback): PointCallback {
         return p => {
-            if(this._isDown) this._upCallback(p)
+            if(this.isPressed) this._upCallback(p)
             cb(p)
             this.isPressed = false
-            this._isDown = false
         }
     }
 
-    private _hoverCallback: PointerCallback
-    private _pressedCallback: PointerCallback
-    private _downCallback: PointerCallback
-    private _upCallback: PointerCallback
-    private _nonAnyInteractiveCallback: PointerCallback
+    private _hoverCallback: PointCallback
+    private _pressedCallback: PointCallback
+    private _downCallback: PointCallback
+    private _upCallback: PointCallback
+    private _nonAnyInteractiveCallback: PointCallback
 
     public isPressed: boolean
-    
-    private _isDown: boolean
     
     constructor(x?: number, y?: number) {
         super(x, y)
 
-        this._isDown = false
         this.isPressed = false
 
         this._downCallback = p => {}
@@ -41,40 +37,39 @@ export default class PointerableObject extends ShapeableObject implements IShape
         this.doIfNotAnyInteracted(p => {})
     }
 
-    get hoverCallback(): PointerCallback {
+    get hoverCallback(): PointCallback {
         return this._hoverCallback
     }
 
-    get pressedCallback(): PointerCallback {
+    get pressedCallback(): PointCallback {
         return this._pressedCallback
     }
 
-    get nonAnyInteractiveCallback(): PointerCallback {
+    get nonAnyInteractiveCallback(): PointCallback {
         return this._nonAnyInteractiveCallback
     }
 
-    doWhenDown(cb: PointerCallback): void {
+    doWhenDown(cb: PointCallback): void {
         this._downCallback = cb
     }
 
-    doWhenUp(cb: PointerCallback): void {
+    doWhenUp(cb: PointCallback): void {
         this._upCallback = cb
     }
 
-    doWhenPressed(cb: PointerCallback): void {
+    doWhenPressed(cb: PointCallback): void {
         this._pressedCallback = p => {
-            if(!this._isDown) this._downCallback(p)
+            if(!this.isPressed) this._downCallback(p)
             cb(p)
             this.isPressed = true
-            this._isDown = true
         }
     }
 
-    doWhenHover(cb: PointerCallback): void {
+    doWhenHover(cb: PointCallback): void {
         this._hoverCallback = this._nonDownCallbackFactory(cb)
     }
 
-    doIfNotAnyInteracted(cb: PointerCallback): void {
+    doIfNotAnyInteracted(cb: PointCallback): void {
         this._nonAnyInteractiveCallback = this._nonDownCallbackFactory(cb)
     }